fb: Add JZ4770 framebuffer driver from Ingenic
Based on driver from linux-2.6.31.3-jz-20110809-FR1.patch.gz.
Only the drivers relevant to JZ4770 are included.
Stripped functionality to what we really need, to make the code more
suitable for restructuring.

Removed all panel support code, this is handled via separate panel
drivers now.

Implemented the check_var function, now SDL applications start up.
This implementation only supports the hardcoded 320x240@16bpp and
320x240@32bpp video modes.

The mmap-ed frame buffer is configured to be fully cacheable. This
improves the raw pixel writing rate by about 200% compared to an
uncached frame buffer. The cache is flushed when a new page starts
being displayed.

Implemented blanking.

Wait-for-vsync is implemented both as an explicit ioctl and implicitly
when changing the vertical panning. The latter allows a more reliable
page flip in combination with a patched SDL.

The driver now always uses the foreground 1 layer in OSD mode. This is
necessary in preparation for TV-out support. Also it will allow use of
the foreground 0 layer for a splash screen and to show OSD messages.

The LCD timing is configured to get a frame rate of exactly 60 Hz.
The panel timings are a bit unconventional, but it works fine with our
panel and it avoids stutters when applications wait both on timers and
on vsync. These timings might not work with TV-out, but we'll deal with
that once TV-out is actually implemented.

The driver now reserves enough memory for triple buffering.

Displays the current logo at the end of the probe function.

Upload frames using the IPU instead of DMA.

The IPU is the Image Processing Unit. It handles scaling with bilinear
or bicubic filtering as well as colorspace conversion.
With this change, the LCD controller does not use the DMA anymore to
upload the framebuffer to the screen; instead, the framebuffer is read
by the IPU, eventually processed, and uploaded directly to the LCD
without crossing the RAM again.

For now, this brand new video driver handles all the features of the
previous one:
- Single buffering with deferred cache flush,
- VSYNC with perfect 60 Hz,
- Double / triple buffering,
- 32 / 16 bpp,
- Blanking, suspend / resume.

Some differences:
- The LCD is always configured for 32 bpp; when a different mode is
  used, the IPU is used to convert the framebuffer to 32 bpp.
- Minor regression: as the IPU takes some time to configure (and will
  take even more when scaling is implemented), garbage is displayed for
  a couple of milliseconds before the first image is displayed.

Now, the framebuffer driver will accept any resolution between 4x4 and
640x480, and will upscale or downscale the output to the panel's
resolution, with a bilinear filter.

By default, black bars are added around the picture in order to keep the
aspect ratio of the original image.
The fullscreen mode, which doesn't respect the aspect ratio, can be
enabled by writing 0 to
/sys/devices/platform/jz-lcd.0/keep_aspect_ratio

This parameter can be switched at runtime, the IPU will be reconfigured
as soon as the parameter is written in sysfs.

When upscaling with an integer ratio (e.g. 3x), we upscale with the
nearest neighbour algorithm, which gives the best results.

Also changed are the EFW/BFW timings to remove the "black line" hack that
were introduced a long time ago for perfect VSYNC when using DMA.

A new parameter can be used to enable/disable downscaling.
Downscaling is disabled by default.
This switch is useful for applications which support higher resolutions
and will use them by default; when downscaling is disabled, it forces
them to render to the panel's native resolution.
The sysfs file is /sys/devices/platform/jz-lcd.0/allow_downscaling

Cut the rightmost column of pixels to avoid showing garbage.
If we are upscaling horizontally, the last columns of pixels
shall be hidden, as they usually contain garbage: the last
resizing coefficients, when applied to the last column of the
input frame, instruct the IPU to blend the pixels with the
ones that correspond to the next column, that is to say the
leftmost column of pixels of the input frame.

Removed manual pin configuration.
We now rely on pinctrl to do this for us.

Switch to sleep pinctrl state when display is blanked.
We don't need a signal on the LCD pins when the panel is off.

Moved headers to driver directory.
Removed panel hooks, use platform-lcd driver instead.
Allow userspace to specify pixclock.

Add support for a display area smaller than the full panel.
This is useful for displays that have overscan, like TVs.

Fix IPU crash occuring with some resolutions.
Some resolutions, like 410x256, would result in 32 coefficients being
written to the scaling table. This causes the IPU to crash, probably
because only 31 slots are available.

Rename jz4760fb -> jzfb, and code cleanup.
This driver is for the jz4770 and it has never been tested on jz4760.
Therefore it doesn't make sense to have references to the jz4760 here.

Remove the global variable jz_panel.
Instead the driver will now use jzfb->panel.
This will make it easier in the future to support multiple panels.

Prepare/unprepare the clock.

Add integer scaling option.

Add device tree support.

Fixed disable timeout.
The previous code would not check the status register after the last
sleep, so it could report a timeout if the status flipped during that
last sleep.

Avoid LCD disable timeout on boot.

Wait until a frame has been sent before registering.
This avoids a white flash on the screen at boot.

Drop legacy code to select LPCLK parent.
Remove useless includes.
Don't hardcode IRQ number.
Removed redundant write of fix.visual.

Properly lock access for par struct and hw regs.

I set myself as the author, since this driver is now very different
from the original driver it was based on, so the original author cannot
be expected to answer questions about it. Proper credit is given in the
comment at the top of the file.

Populate nested devices
This allows the LCD driver to be a child of the LCDC driver, which
makes it easy to match the LCD with the right framebuffer.