ASoC: implement a driver for the jz4780 & CI20
This patch adds an ALSA SoC audio driver for the CI20 & the Ingenic
jz4780 SoC that it's built around. Both the internal codec headphone
output & the SoCs HDMI output are supported. The hardware only allows
for one of the headphone or HDMI outputs to be active at any one time
since they are both driven through the AIC block with different setups
(primarily bit clock source, direction). The output to use may be
selected by toggling the 'HDMI' kcontrol which is exposed by ALSA.

Input through a microphone connected as part of a 4-pin OMTP-compliant
headset is also supported, but not whilst outputting audio via HDMI
due to the aforementioned hardware limitations.

Tested with various combinations of VLC, Gnome sound recorder &
pulseaudio.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>