jzdma: workaround to support interrupts for cyclic DMA
The DMA controller found in the Ingenic JZ4780 SoC does not support
firing an interrupt after processing each descriptor, only after
processing an entire terminated list of descriptors. This patch adds a
workaround to support this functionality by degrading from a cyclic set
of descriptors to a set of individual descriptors which are submitted
in sequence, each in response to the completion of the prior descriptor.

This will be used in a following patch to add an ASoC audio driver.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>