CI20: add a keymap for the jz_remote IR driver
Commit b5cf574 "make  ingenic remote works  better  on hdmi80" added a
key_maps property to the jz_remote_board_data struct. The jz_remote
driver expects this to point to an array of 256 integers mapping from
IR codes as the index to KEY_ values. The jz_remote driver does not
check whether the key_maps field is set, so if key_maps is left set to
NULL then it will be dereferenced & trigger a kernel panic when the
jz_remote driver is probed.

This commit adds a key map to the CI20 platform data which matches
the key mapping which was hardcoded in the jz_remote driver prior to
commit b5cf574.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>