jzdma: use spin_lock_irqsave variant in place of spin_lock_bh
The ALSA driver will need to terminate DMA transactions for a channel in
the context of a syscall. This results in taking the channel lock in
jzdma_terminate_all which was done using spin_lock_bh since the lock may
also be taken in one of a number of tasklets. However this is not valid
within the context of a syscall since interrupts are disabled triggering
a WARN_ON_ONCE in _local_bh_enable_ip. Switching to the _irqsave makes
terminating DMA transactions safe for all uses. The same is done for DMA
descriptor submission.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>