jz4780/bch: fix the reported count of corrected errors
The bch_request_t cnt_ecc_errors field is supposed to be a count of ECC
errors corrected by a BCH_REQ_{DECODE_,}CORRECT operation. It is
reported directly to the MTD layer by drivers/mtd/nand/jz4780_nand.c as
the number of bits flipped by ECC correction, and as such used by layers
such as UBI to approximate whether a region of flash is in imminent
danger of failure.

The value it should contain can be read directly from TERRC in the BHINT
register, however the value extracted was not being shifted from its
position in the register. This led to a single bit error being reported
to MTD as 65537 bit errors, for example. Correct the count by shifting
the value appropriately.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>