jz4780/bch: fix the BCH ECC correction code
I have no idea what the previous code was attempting to do, it made no
sense and mangled the wrong bits of data. This patch corrects the code
to actually match up with what the hardware provides and to flip the
right bits of output data. Verified by booting an UBIFS root
filesystem, modifying various files, rebooting and repeating a few
times.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>