CI20: switch to hardware BCH error code correction for NAND
U-boot requires the hardware BCH block because software BCH is too large
to fit into SPL. If U-boot & Linux are going to access the same data
then they need to use the same ECC algorithm. I'm told that the hardware
BCH block should not be slower than software despite comments in other
board setup files (though still need to verify that). So now that the
hardware BCH code is working correctly following the prior few commits,
make use of it for the CI20.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>