CI20: rename "root" mtd partition to "system"
This partition will by default be accessed with UBI, and may contain UBI
volumes other than "root". For example U-boot will by default expect the
kernel to be present in an UBI volume named "boot". So rename "root" to
make it clearer that it doesn't have to only contain the root
filesystem, but may contain an arbitrary number of UBI volumes. This
also matches the name U-boot uses for the partition.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>