MIPS: page.h: revert change to VM_DATA_DEFAULT_FLAGS
This reverts the following commits:
ba0dc0e9: "fix cts bug testNoExecutePermission"
26779445: "fix build error"

As of Android CTS revision 1235d085, the NX test should not fail on
MIPS, making this change unnecessary. It also prevents ld.so from using
mprotect() to make the stack executable when a library requires it,
which breaks Iceweasel in Debian 7.

Signed-off-by: Alex Smith <alex.smith@imgtec.com>