jz4780_camera: various fixes to DMA setup
A call to buf_queue should not initiate DMA for all queued buffers,
only the first buffer in the queue, nor should it assume that all
queued buffers are contiguous in physical memory (memory within
each buffer is contiguous, but buffers do not have to be contiguous
with each other). Modified to instead only set up DMA for a buffer
when it becomes active, and use the correct physical addresses. This
fixes a kernel panic when using V4L2_MEMORY_MMAP buffers due to
memory corruption.

Additionally, temporarily disable the CIM/DMA while setting up a
new DMA transfer, as well as when there are no more queued buffers.
This fixes repeated "DMA End IRQ with no active buffer" messages in
the kernel log.

Finally, when deactivating the device, the CIM should be disabled
before disabling its clocks.

Signed-off-by: Alex Smith <alex.smith@imgtec.com>