MIPS: jz4780: give UART4 precedence over LCD in GPIO setup
UART4 shares pins with the LCDC. The LCD pins are always reconfigured
regardless of any Kconfig settings, and as the LCD pin configuration was
specified after UART4 in the configuration array, the configuration for
UART4 would always be overridden making it unusable. Move the UART4
configuration after the LCD in the array so that it will take precedence
when it is enabled in Kconfig.

Signed-off-by: Alex Smith <alex.smith@imgtec.com>