sound/soc: ci20: fix headset type detection
The TS3A226AE headset type detection switch requires a rising edge on its
enable pin to trigger the detection sequence. We can use the headphone
detect GPIO to trigger this. Override the jack_status_check callback for
the GPIO to toggle the enable pin, which causes a rising edge whenever
something is connected to the headphone jack. Also increased the debounce
time to give a little more time for the connector to be fully inserted
before triggering the detection, and moved the gpio_request calls to
before the platform device is registered to ensure that they have been
requested by the time the ALSA device is initialized.

Signed-off-by: Alex Smith <alex.smith@imgtec.com>