CI20: updates for IW8103 WiFi on rev 1.2
Set the WLAN_PW_EN line high when powering up the WiFi to make the IW8103
on rev 1.2 boards work correctly, and change the default configuration to
use the right firmware/NVRAM path for the 8103. Technically, WLAN_PW_EN
should have needed to be set high on rev 1.1 as well, but for some reason
worked without. With this change rev 1.1 still works correctly.

Signed-off-by: Alex Smith <alex.smith@imgtec.com>