CI20: Generate MAC from chip ID if no address programmed in EFUSE
If there is not a valid MAC address programmed in the EFUSE (i.e. on
pre-production boards), generate a locally-administered address from
the unique parts of the chip ID stored in EFUSE instead. This matches
the behaviour of U-Boot, and gives us fixed MAC addresses on the pre-
production boards, rather than randomly-generating one each boot.

Signed-off-by: Alex Smith <alex.smith@imgtec.com>