MIPS: Add clock_ignore_unused to ci20_defconfig
Without the clock_ignore_unused parameter on the kernel command line,
the kernel will disable some necessary clocks & cause the system to
hang during boot. This parameter is currently hardcoded in the bootargs
provided by the chosen node of the DT, but that will shortly be removed
in favor of honoring the command line provided by U-boot.

Add clock_ignore_unused to CONFIG_CMDLINE such that it is appended to
the command line regardless of what the user configures U-boot to
provide. This ensures that kernels built from the defconfig will boot
regardless of the version of U-boot, and allows for the parameter to
be removed easily in the future once the clock code is able to keep the
required clocks enabled.

[paul.burton@imgtec.com: Rewrote commit message]

Signed-off-by: Zubair Lutfullah Kakakhel <Zubair.Kakakhel@imgtec.com>
Signed-off-by: Paul Burton <paul.burton@imgtec.com>