Upgrade ion/dma-buf to latest version.
This again imports changes that have already been made upstream in
android-3.4. It moves the dma-buf and ion features up to the versions
in android-3.4, which is required by the PowerVR SGX driver.

Again it would be better if android-3.4 was merged up rather than me
cherry-picking these files from it.

Note that this change does not add the changes necessary to build
the subsystem on MIPS; that will be done in a follow-up commit.