sound: soc: jz4740: Fix broken audio for !mono sound
The jz4780 SoC offers the possibility to configure the number of
channels used for playback. Right now, the corresponding bits in the
configuration register are always zero, which means that the hardware is
only configured for monaural sound.

With this commit, the hardware is now correctly configured for the
number of channels requested by the ALSA core are.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>