IIO: Ingenic JZ47xx: Avoid writing non-existent register on JZ4740
The ADCLK register for configuring clock dividers doesn't exist on
JZ4740, so we should avoid writing it there.

I introduced a function rather than a flag because there is a lot
of variation between the ADCLK registers on JZ47xx SoCs, both in
the internal layout of the register and in the frequency range
supported by the SADC. So this solution should make it easier
to add support for other JZ47xx SoCs later.