MIPS: ingenic: Add support for the YLM RG-350(M), RG-280{M,V}
Add support for the YLM (known as the Anbernic in the western
world) RG-280V, RG-280M, RG-350 and RG-350M.

The YLM RG-280V is almost a carbon copy of the GCW-Zero, with a
different LCD panel (320x480 non-square pixels in diamond pattern),
L2/R2 and Vol +/- buttons added, and with the analog stick, ACT8600
regulator, RDA5807 radio, MXC6225 accelerometer, ITE6610 HDMI chips
removed.

The YLM RG-280M is a RG-280V with a clickable analog stick and a
ITE66121 HDMI chip added.

The YLM RG-350 is basically a RG-280M, with a proper 320x240 IPS screen
this time, and a second clickable analog stick.

The RG-350M is a slightly modified version of the RG-350, the only
hardware difference being the screen with a 640x480 resolution vs.
320x240 for the non-M variant.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>