gpu/drm: ingenic: Add trick to support 16bpp on 24-bit panels
If the panel interface is 24-bit but our primary plane is 16bpp,
configure as if the panel was 18-bit. This tricks permits the display
of 16bpp data on a 24-bit panel by wiring each color component to the
MSBs of the 24-bit interface.

v2: Check bytes-per-pixel count instead of fourcc format

Signed-off-by: Paul Cercueil <paul@crapouillou.net>