gpu/drm: Ingenic: Add support for the IPU
Add support for the Image Processing Unit (IPU) found in all Ingenic
SoCs.

The IPU can upscale and downscale a source frame of arbitrary size
ranging from 4x4 to 4096x4096 on newer SoCs, with bicubic filtering
on newer SoCs, bilinear filtering on older SoCs. Nearest-neighbour can
also be obtained with proper coefficients.

It can also convert between many different YUV and RGB pixel formats.

Starting from the JZ4760, the IPU supports a mode where its output is
sent directly to the LCDC, without having to be written to RAM first.
This makes it possible to use the IPU as the primary plane on the
compatible SoCs, and have it convert and scale anything the userspace
asks for to what's available for the display.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>