dt-bindings: timer: ingenic: Support watchdog/PWM on JZ4725B SoC
Add compatible strings to the watchdog and PWM nodes to support the
Ingenic JZ4725B SoC.

The watchdog on the JZ4725B is similar  to the one on the JZ4740, so the
new ingenic,jz4725b-watchdog requires a fallback to
ingenic,jz4740-watchdog.

The PWM on the JZ4725B has only 6 channels, instead of 8 for the JZ4740,
so a new ingenic,jz4725b-pwm compatible string has been introduced.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>