iio/adc: ingenic: Program ADCMD register on compatible SoCs
JZ4770 and later SoCs have a neat command feature. Up to 32 commands can
be programmed, each one corresponding to a sampling job. What makes it
interesting, is that it is very low-level. It is possible to change the
low-voltage reference, the high-voltage reference, have them connected
to VCC, GND, or one of the X-/X+ or Y-/Y+ pins.

Right now, with this patch we use the ADCMD feature when available to
sample the X+/Y+ pins relative to GND, which is the same behaviour as
before. Later, new channels could be added to allow sampling the X-/Y-
pins, or to sample X/Y in differential mode.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>