dt-bindings: display: Add ABT Y030XX067A panel bindings
The Asia Better Technology (ABT) Y030XX067A panel is a 3.0" 320x480
24-bit IPS LCD panel. Its particularity is that it has non-square pixels
(as it is 4:3 for a resolution of 320x480), and that it requires odd
lines to be sent as RGB and even lines to be sent as GRB on its 8-bit
bus.

v2: Add missing 'required' and 'additionalProperties' properties.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>