rtc: jz4740: Unconditionally enable CLK32K pin on JZ4780
On JZ4780, the CLK32K pin is configurable. By default, it is configured
as a GPIO in input mode, and its value can be read through GPIO PD14.

Since it is very unlikely that anybody will need to use this pin as
anything else than for the 32 kHz clock, hardcode this configuration.

This permits the WiFi/Bluetooth chip to work on the MIPS CI20 board,
which does source one of its clocks from the CLK32K pin.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>